+++ test/test_decoders/dual_wheel/dual_wheel.cpp
@@ -81,7 +81,7 @@ static void test_dualwheel_newIgn_12_1()
 static void test_getCrankAngle(void)
 {
   extern decoder_status_t decoderStatus;
-  extern volatile unsigned long toothLastToothTime;
+  extern volatile uint32_t toothLastToothTime;
   extern volatile int toothCurrentCount;
   extern volatile bool revolutionOne;
